+++ addons/compat_ws/compat_ws_ballistics/CfgMagazineWells.hpp
@@ -1,6 +1,4 @@
-
 class CfgMagazineWells {
-
     class AA40_12Gauge { // Weapon initSpeed -1: ACE_barrelLength 762mm (12 Gauge SAAMI test barrel 30"), ACE_muzzleVelocities 610m/s (2000fps ICAO)
         ADDON[] = { // standart version 18" 457.2mm initSpeed -0.940574 574m/s according to ACE_muzzleVelocities ICAO
             "ACE_20Rnd_12Gauge_AA40_Slug_lxWS",
