+++ addons/compat_ws/compat_ws_ballistics/config.cpp
@@ -1,22 +1,23 @@
 #include "script_component.hpp"
 
 class CfgPatches {
-    class ADDON {
+    class SUBADDON {
         name = COMPONENT_NAME;
         units[] = {};
         weapons[] = {};
         requiredVersion = REQUIRED_VERSION;
-        requiredAddons[] = {"data_f_lxWS_Loadorder", "ace_common", "ace_ballistics"};
+        requiredAddons[] = {
+            "ace_ballistics",
+            QUOTE(ADDON)
+        };
         skipWhenMissingDependencies = 1;
         author = ECSTRING(common,ACETeam);
-        authors[] = {"Mike"};
+        authors[] = {"Laid3acK"};
         url = ECSTRING(main,URL);
         VERSION_CONFIG;
+        addonRootClass = QUOTE(ADDON);
     };
 };
 
-#include "CfgEventHandlers.hpp"
-#include "CfgVehicles.hpp"
-#include "CfgWeapons.hpp"
-#include "CfgAmmo.hpp"
 #include "CfgMagazines.hpp"
+#include "CfgMagazineWells.hpp"
