+++ addons/compat_ws/compat_ws_ballistics/script_component.hpp
@@ -1,0 +1,3 @@
+#define SUBCOMPONENT ballistics
+#define SUBCOMPONENT_BEAUTIFIED Ballistics
+#include "..\script_component.hpp"
